Handle "same-scheme" URLs
xhr-xdr-adapter throws an exception if it receives a "same-scheme" URL, i.e., one that is relative with respect to protocol (of the form `//example.com`). This pull request resolves the exception by prepending location.protocol to remoteUrl. (NB: location.protocol already has the trailing colon, so no need to add that).